# Avoid extra allocations in YamlScalarNode GetHashCode