--- a/YamlDotNet/Core/HashCode.cs
+++ b/YamlDotNet/Core/HashCode.cs
@@ -42,16 +42,6 @@ namespace YamlDotNet.Core
             return CombineHashCodes(h1, GetHashCode(o2));
         }
 
-        public static int CombineHashCodes(object? first, params object?[] others)
-        {
-            var hashCode = GetHashCode(first);
-            foreach (var other in others)
-            {
-                hashCode = CombineHashCodes(hashCode, other);
-            }
-            return hashCode;
-        }
-
         private static int GetHashCode(object? obj)
         {
             return obj != null ? obj.GetHashCode() : 0;