@aaubry 2022-07-23
This feature has been released in [version 12.0.0](https://github.com/aaubry/YamlDotNet/releases/tag/v12.0.0).